App provides defaults for getStateType and init
This is to accommodate file-based hot-swapping in the Verse editor.
The hot-swapper refuses to overwrite properties on the window that
are already defined, so the real definitions of init and getStateType
must be the first ones on the window.

This commit also removes waitForInput; it's a naive implementation and
apps can define their own for now.